+++ packages/orq-rc/src/models/components/evaluatorresponseragas.ts
@@ -50,8 +50,8 @@ export const EvaluatorResponseRagas$inboundSchema: z.ZodType<
 > = z.object({
   _id: z.string(),
   description: z.string(),
-  created: z.string().default("2026-07-03T12:12:55.302Z"),
-  updated: z.string().default("2026-07-03T12:12:55.303Z"),
+  created: z.string().default("2026-07-03T14:05:30.205Z"),
+  updated: z.string().default("2026-07-03T14:05:30.205Z"),
   updated_by_id: z.nullable(z.string()).optional(),
   guardrail_config: z.any().optional(),
   type: z.literal("ragas"),