+++ packages/orq-rc/src/models/operations/postv2feedbackevaluation.ts
@@ -193,7 +193,7 @@ export const RequestBody3$outboundSchema: z.ZodType<
   source: PostV2FeedbackEvaluationRequestBodyFeedbackSource$outboundSchema
     .default("orq"),
   explanation: z.string().optional(),
-  reviewedAt: z.date().default(() => new Date("2026-07-03T12:13:11.219Z"))
+  reviewedAt: z.date().default(() => new Date("2026-07-03T14:05:50.355Z"))
     .transform(v => v.toISOString()),
   type: z.literal("string_array"),
   values: z.array(z.string()),
@@ -247,7 +247,7 @@ export const PostV2FeedbackEvaluationRequestBody2$outboundSchema: z.ZodType<
     "orq",
   ),
   explanation: z.string().optional(),
-  reviewedAt: z.date().default(() => new Date("2026-07-03T12:13:11.218Z"))
+  reviewedAt: z.date().default(() => new Date("2026-07-03T14:05:50.354Z"))
     .transform(v => v.toISOString()),
   type: z.literal("number"),
   value: z.number(),
@@ -305,7 +305,7 @@ export const PostV2FeedbackEvaluationRequestBody1$outboundSchema: z.ZodType<
   humanReviewId: z.string().optional(),
   source: RequestBodySource$outboundSchema.default("orq"),
   explanation: z.string().optional(),
-  reviewedAt: z.date().default(() => new Date("2026-07-03T12:13:11.217Z"))
+  reviewedAt: z.date().default(() => new Date("2026-07-03T14:05:50.353Z"))
     .transform(v => v.toISOString()),
   type: z.literal("string"),
   value: z.string(),