+++ packages/orq-rc/RELEASES.md
@@ -10549,3 +10549,13 @@ Based on:
 - [typescript v4.11.0-rc.77] packages/orq-rc
 ### Releases
 - [NPM v4.11.0-rc.77] https://www.npmjs.com/package/@orq-ai/node/v/4.11.0-rc.77 - packages/orq-rc
+
+## 2026-07-03 14:08:21
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.788.0 (2.915.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.11.0-rc.78] packages/orq-rc
+### Releases
+- [NPM v4.11.0-rc.78] https://www.npmjs.com/package/@orq-ai/node/v/4.11.0-rc.78 - packages/orq-rc